backend_flask/requirements.txt
flask_cors
Flask
python-dotenv
google-generativeai